tests/requirements.txt
mock
pytest-cov
requests
